# doc/diode-server-api-spec(swagger resolved).yaml
"openapi": "3.1.0"
"info":
  "title": "Спецификация API сервера информации об объектах инфраструктуры"
  "description": "Имитирует запросы для получения данных объектов инфраструктуры из базы паспортных данных ЕК АСУИ"
  "version": "1.0"
"servers":
  -
    "description": "diode-server default url"
    "url": "http://localhost:8070"
...
"paths":
...
  "/api/v1/objinfo/p/joint/{joint_id}":
    "get":
      "parameters":
        -
          "description": "Идентификатор рельсового стыка в базе данных"
          "in": "path"
          "name": "joint_id"
          "schema":
            "type": "integer"
            "format": "int64"
            "examples":
              - 1234
          "required": true
        -
          "description": "Тип пути, на котором расположен объект."
          "in": "query"
          "name": "waytype"
          "schema":
            "type": "string"
            "description": "Тип пути"
            "enum":
              - "INTERSTATION_TRACK"
              - "STATION_MAIN_WAY"
              - "STATION_WAY"
            "examples":
              - "STATION_WAY"
          "required": true
        -
          "description": "Тип стыка."
          "in": "query"
          "name": "jointtype"
          "schema":
            "type": "string"
            "examples":
              - "CONDUCTING"
            "enum":
              - "CONDUCTING"
              - "ISOLATING_SIGNAL"
              - "ISOLATING_SWITCH"
          "required": true
      "description": "Возвращает заготовку DTO с исчерпывающей информацией, доступной в базе данных о рельсовом стыке для дальнейшего создания справки-характеристики о нем. Дополнительное поле возвращаемого DTO равно null, так как не относятся к сервису (заполняются пользователем в UI). Возвращает 404 ошибку, если передан неверный id стыка (не найден в БД), 400 ошибку, если указан неправильный параметр."
      "operationId": "getJointInfoDto"
      "responses":
        "200":
          "content":
            "application/json":
              "schema":
                "description": "DTO со всей необходимой информацией для создания характеристики рельсового стыка."
                "properties":
                  "locationDescription":
                    "type": "string"
                    "description": "Строка с названием конкретной локации."
                    "examples":
                      - "Где-нибудьск"
                  "jointLocation":
                    "type": "object"
                    "properties":
                      "wayType":
                        "type": "string"
                        "enum":
                          - "INTERSTATION_TRACK"
                          - "STATION_MAIN_WAY"
                          - "STATION_WAY"
                        "description": "Тип пути (станционный или главный)"
                        "examples":
                          - "STATION_WAY"
                      "locationId":
                        "type": "integer"
                        "format": "int64"
                        "description": "Идентификатор объекта-локации, в пределах которого находится объект"
                        "examples":
                          - 8
                      "wayNumber":
                        "type": "string"
                        "description": "Строка, содержащая название (номер) пути"
                        "examples":
                          - "5"
                      "km":
                        "type": "integer"
                        "description": "Целочисленное значение километра, на котором находится железнодорожный объект"
                        "examples":
                          - 0
                      "picket":
                        "type": "integer"
                        "description": "Пикет километра, на котором находится железнодорожный объект"
                        "examples":
                          - 1
                      "meter":
                        "type": "integer"
                        "description": "Метр пикета, на котором находится железнодорожный объект"
                        "examples":
                          - 1
                      "lineSide":
                        "type": "string"
                        "description": "Сторонность рельсовой нити"
                        "enum":
                          - "LEFT"
                          - "RIGHT"
                        "examples":
                          - "LEFT"
                    "description": "DTO с информацией о расположении стыка из базы паспортных данных."
                    "examples":
                      -
                        "wayType": "STATION_WAY"
                        "locationId": 8
                        "wayNumber": 3
                        "km": "0,"
                        "picket": "7,"
                        "meter": "57,"
                        "lineSide": "RIGHT"
                  "wayCharacteristics":
                    "type": "object"
                    "properties":
                      "slope":
                        "type": "number"
                        "format": "float"
                        "description": "Уклон"
                        "examples":
                          - -1.5
                      "straight":
                        "type": "boolean"
                        "description": "Флаг, определяющий, прямой участок или кривая"
                        "examples":
                          - false
                      "curveSide":
                        "type": "string"
                        "description": "Сторонность кривой (null для прямого участка)"
                        "examples":
                          - "ЛЕВАЯ"
                      "radius":
                        "type": "integer"
                        "description": "Радиус кривой (null для прямого участка)"
                        "examples":
                          - 1000
                      "railElevation":
                        "type": "integer"
                        "description": "Возвышение рельса (null для прямого участка)"
                        "examples":
                          - 135
                      "passengerTrainSpeed":
                        "type": "integer"
                        "description": "Скорость движения пассажирских поездов"
                        "examples":
                          - 120
                      "freightTrainSpeed":
                        "type": "integer"
                        "description": "Скорость движения грузовых поездов"
                        "examples":
                          - 80
                      "passedTonnageBeforeInstall":
                        "type": "number"
                        "format": "float"
                        "description": "Пропущенный тоннаж до укладки"
                        "examples":
                          - 500.75
                      "passedTonnageAfterInstall":
                        "type": "number"
                        "format": "float"
                        "description": "Пропущенный тоннаж после укладки"
                        "examples":
                          - 200.57
                      "freightIntensity":
                        "type": "number"
                        "format": "float"
                        "description": "Грузонапряженность"
                        "examples":
                          - 95.57
                      "lineClassSpecialization":
                        "type": "string"
                        "description": "Класс и специализация линии"
                        "examples":
                          - "1Т"
                      "groupClassCode":
                        "type": "string"
                        "description": "Код группы пути"
                        "examples":
                          - "1О"
                      "railType":
                        "type": "string"
                        "description": "Тип рельсов"
                        "examples":
                          - "Р65"
                      "railCategory":
                        "type": "string"
                        "description": "Категория рельсов"
                        "examples":
                          - "Т1"
                      "thermalHardening":
                        "type": "string"
                        "description": "Тип термоупрочнения рельсов"
                        "examples":
                          - "ДИФФЕРЕНЦИАЛЬНО-ТЕРМОУПРОЧНЕННЫЕ"
                      "wayType":
                        "type": "string"
                        "description": "Тип пути"
                        "examples":
                          - "БЕССТЫКОВОЙ"
                      "factory":
                        "type": "string"
                        "description": "Завод-изготовитель рельсов"
                        "examples":
                          - "К"
                      "factoryYear":
                        "type": "string"
                        "description": "Год производства рельсов"
                        "examples":
                          - "2015"
                      "installationDate":
                        "type": "string"
                        "format": "date"
                        "description": "Дата укладки рельсов в путь"
                        "examples":
                          - "01-07-2021"
                      "sleeperMaterial":
                        "type": "string"
                        "description": "Материал шпал"
                        "examples":
                          - "Ж/Б"
                      "sleepersPerKm":
                        "type": "integer"
                        "description": "Количество шпал на км"
                        "examples":
                          - 1943
                      "fasteningType":
                        "type": "string"
                        "description": "Тип рельсовых скреплений"
                        "examples":
                          - "АРС"
                      "ballastType":
                        "type": "string"
                        "description": "Тип балласта"
                        "examples":
                          - "ЩЕБЕНЬ"
                      "ballastHeight":
                        "type": "integer"
                        "description": "Высота балласта"
                        "examples":
                          - 45
                      "capitalRepairDate":
                        "type": "string"
                        "format": "date"
                        "description": "Год и месяц капитального ремонта"
                        "examples":
                          - "01-07-2021"
                      "capitalRepairType":
                        "type": "string"
                        "description": "Тип капитального ремонта"
                        "examples":
                          - "КРС"
                      "intermediateRepairDate":
                        "type": "string"
                        "format": "date"
                        "description": "Год и месяц промежуточного ремонта"
                        "examples":
                          - "01-09-2024"
                      "intermediateRepairType":
                        "type": "string"
                        "description": "Тип промежуточного ремонта"
                        "examples":
                          - "ППВ"
                    "description": "DTO с информацией о характеристиках пути в конкретном месте."
                  "joint":
                    "type": "object"
                    "properties":
                      "id":
                        "type": "integer"
                        "format": "int64"
                        "description": "Идентификатор стыка в БД."
                        "examples":
                          - 4321
                      "lineSide":
                        "type": "string"
                        "description": "Сторона рельсовой нити, на которой расположен стык."
                        "examples":
                          - "ПРАВАЯ"
                      "status":
                        "type": "string"
                        "description": "Статус рельсового стыка."
                        "examples":
                          - "АКТИВЕН"
                      "padsType":
                        "type": "string"
                        "description": "Тип накладок."
                        "examples":
                          - "ДВУХГОЛОВЫЕ"
                      "padAmountOfHoles":
                        "type": "integer"
                        "description": "Количество болтовых отверстий."
                        "examples":
                          - 6
                      "gap":
                        "type": "integer"
                        "description": "Значение величины стыкового зазора."
                        "examples":
                          - 10
                      "verticalStep":
                        "type": "number"
                        "format": "float"
                        "description": "Значение величины вертикальной ступеньки."
                        "examples":
                          - 0.5
                      "horizontalStep":
                        "type": "number"
                        "format": "float"
                        "description": "Значение величины горизонтальной ступеньки."
                        "examples":
                          - 0.9
                      "lastMeasureDate":
                        "type": "string"
                        "format": "date"
                        "description": "Дата последнего измерения характеристик рельсового стыка."
                        "examples":
                          - "01-10-2024"
                      "lastMeasureRailTemp":
                        "type": "integer"
                        "description": "Температура рельса в момент последнего измерения характеристик."
                        "examples":
                          - 25
                      "contactResistance":
                        "type": "integer"
                        "description": "Величина переходного сопротивления в токопроводящем рельсовом стыке."
                        "examples":
                          - 200
                      "lastResistanceMeasureDate":
                        "type": "string"
                        "format": "date"
                        "description": "Дата последнего измерения переходного сопротивления в токопроводящем стыке."
                        "examples":
                          - "15-08-2024"
                      "lastPadsRemovalDate":
                        "type": "string"
                        "format": "date"
                        "description": "Дата последнего снятия накладок токопроводящего стыка."
                        "examples":
                          - "01-04-2024"
                      "connectorTypes":
                        "type": "string"
                        "description": "Типы рельсовых соединителей, установленных в токопроводящем рельсовом стыке."
                        "examples":
                          - "СРСП"
                      "connectorInstallationDate":
                        "type": "string"
                        "format": "date"
                        "description": "Дата установки рельсового соединителя(-лей) в токопроводящем рельсовом стыке."
                        "examples":
                          - "10-10-2022"
                      "stationWayPlace":
                        "type": "object"
                        "properties":
                          "stationWay":
                            "properties":
                              "id":
                                "type": "integer"
                                "format": "int64"
                                "description": "Идентификатор станционного пути"
                                "examples":
                                  - 7
                              "number":
                                "type": "string"
                                "description": "Номер (название) станционного пути"
                                "examples":
                                  - "Путь 1A"
                              "station":
                                "properties":
                                  "id":
                                    "type": "integer"
                                    "format": "int64"
                                    "description": "Идентификатор станции (структурного предприятия)"
                                    "examples":
                                      - 1001
                                  "name":
                                    "type": "string"
                                    "description": "Название станции (структурного предприятия)"
                                    "examples":
                                      - "Где-нибудьск"
                                "type": "object"
                                "description": "Мини-DTO сущности железнодорожной станции"
                            "type": "object"
                            "description": "Мини-DTO сущности станционного пути железнодорожной станции"
                          "meterOnStationWay":
                            "type": "integer"
                            "description": "Метр станционного пути железнодорожной станции, на котором находится описываемое место"
                            "examples":
                              - 75
                        "description": "DTO с информацией о конкретном месте станционного пути."
                    "description": "DTO с информацией о самом стыке."
                  "metaData":
                    "examples":
                      - null
                "type": "object"
          "description": "Все необходимые данные для формирования JointInfoDto получены без выброса исключений"
        "500":
          "content":
            "application/json":
              "schema":
                "type": "object"
                "description": "DTO, содержащий данные о возникшем в ходе работы программы исключении"
                "properties":
                  "timestamp":
                    "type": "string"
                    "description": "Дата и время когда произошло исключение (в формате \"dd-MM-yyyy HH:mm:ss.SS\")"
                    "examples":
                      - "27-07-2024 07:07:07.77"
                  "exception":
                    "type": "string"
                    "description": "Наименование исключения"
                    "examples":
                      - "NullPointerException"
                  "message":
                    "type": "string"
                    "description": "Сообщение, с которым было выброшено исключение"
                    "examples":
                      - "Не указан обязательный параметр при запросе данных"
                  "details":
                    "type": "string"
                    "description": "Дополнительная информация, касающаяся исключения"
                    "examples":
                      - "stacktrace"
              "example":
                "timestamp": "27-07-2024 09:10:50.77"
                "exception": "DatabaseConnectionException"
                "message": "Failed to connect to database"
                "details": "... stacktrace ..."
          "description": "Произошла ошибка на сервере во время операции"
        "400":
          "content":
            "application/json":
              "schema":
                "type": "object"
                "description": "DTO, содержащий данные о возникшем в ходе работы программы исключении"
                "properties":
                  "timestamp":
                    "type": "string"
                    "description": "Дата и время когда произошло исключение (в формате \"dd-MM-yyyy HH:mm:ss.SS\")"
                    "examples":
                      - "27-07-2024 07:07:07.77"
                  "exception":
                    "type": "string"
                    "description": "Наименование исключения"
                    "examples":
                      - "NullPointerException"
                  "message":
                    "type": "string"
                    "description": "Сообщение, с которым было выброшено исключение"
                    "examples":
                      - "Не указан обязательный параметр при запросе данных"
                  "details":
                    "type": "string"
                    "description": "Дополнительная информация, касающаяся исключения"
                    "examples":
                      - "stacktrace"
              "example":
                "timestamp": "27-07-2024 09:10:50.77"
                "exception": "IllegalArgumentException"
                "message": "Параметр jointtype не может быть null."
                "details": "... stacktrace ..."
          "description": "Исключение возникло из-за ошибки в запросе."
        "404":
          "content":
            "application/json":
              "schema":
                "type": "object"
                "description": "DTO, содержащий данные о возникшем в ходе работы программы исключении"
                "properties":
                  "timestamp":
                    "type": "string"
                    "description": "Дата и время когда произошло исключение (в формате \"dd-MM-yyyy HH:mm:ss.SS\")"
                    "examples":
                      - "27-07-2024 07:07:07.77"
                  "exception":
                    "type": "string"
                    "description": "Наименование исключения"
                    "examples":
                      - "NullPointerException"
                  "message":
                    "type": "string"
                    "description": "Сообщение, с которым было выброшено исключение"
                    "examples":
                      - "Не указан обязательный параметр при запросе данных"
                  "details":
                    "type": "string"
                    "description": "Дополнительная информация, касающаяся исключения"
                    "examples":
                      - "stacktrace"
              "example":
                "timestamp": "27-07-2024 09:10:50.77"
                "exception": "ObjectNotFoundException"
                "message": "Не удалось найти в базе данных объект токопроводящий рельсовый стык на станционном пути с ид=-1"
                "details": "... stacktrace ..."
          "description": "Отсутствуют данные о запрашиваемом рельсовом стыке"
      "summary": "Получение DTO с исчерпывающими данными, необходимыми для создания справки-характеристики рельсового стыка."
      "tags":
        - "Путейцы: справки-характеристики объектов"
...